+++ lib/prompt/io/confirm.ex
@@ -3,7 +3,8 @@ defmodule Prompt.IO.Confirm do
 
   alias __MODULE__
   alias Prompt.IO.Color
-  import IO, only: [read: 2]
+  alias IO.ANSI
+  import IO, only: [write: 1, read: 2]
 
   @type t :: %Confirm{
           color: Color.t(),
@@ -13,8 +14,7 @@ defmodule Prompt.IO.Confirm do
           question: binary(),
           error: nil | binary(),
           answer: nil | binary(),
-          mask_line: boolean(),
-          displayfn: (String.t() -> :ok)
+          mask_line: boolean()
         }
 
   defstruct [
@@ -25,29 +25,47 @@ defmodule Prompt.IO.Confirm do
     :answer,
     :error,
     :default_answer,
-    :mask_line,
-    :displayfn
+    :mask_line
   ]
 
   @spec new(binary(), keyword()) :: t()
-  def new(question, options, displayfn \\ &IO.write(&1)) do
+  def new(question, options) do
     %Confirm{
-      color: Keyword.get(options, :color),
-      background_color: Keyword.get(options, :background_color),
+      color: Keyword.get(options, :color, ANSI.default_color()),
       trim: Keyword.get(options, :trim),
       default_answer: Keyword.get(options, :default_answer),
       question: question,
       answer: nil,
       error: nil,
-      mask_line: Keyword.get(options, :mask_line, false),
-      displayfn: displayfn
+      mask_line: Keyword.get(options, :mask_line, false)
     }
   end
 
   defimpl Prompt.IO do
-    @spec display(Prompt.IO.Confirm.t()) :: Prompt.IO.Confirm.t()
+    @spec display(Confirm.t()) :: Confirm.t()
     def display(%Confirm{} = confirm) do
-      :ok = confirm.displayfn.("#{confirm.question} #{confirm_text(confirm.default_answer)} ")
+      if confirm.mask_line do
+        ANSI.format([
+          :reset,
+          background_color(confirm),
+          confirm.color,
+          "#{confirm.question} #{confirm_text(confirm.default_answer)} ",
+          :reset
+        ])
+        |> write()
+      else
+        [
+          :reset,
+          background_color(confirm),
+          confirm.color,
+          "#{confirm.question} #{confirm_text(confirm.default_answer)} ",
+          :reset,
+          without_newline(confirm.trim)
+        ]
+        |> ANSI.format()
+        |> write()
+      end
+
       confirm
     end
 
@@ -83,5 +101,15 @@ defmodule Prompt.IO.Confirm do
     defp _evaluate_confirm("n", _), do: :no
     defp _evaluate_confirm("", confirm), do: confirm.default_answer
     defp _evaluate_confirm(_, confirm), do: confirm |> display() |> evaluate()
+
+    defp background_color(display) do
+      case display.background_color do
+        nil -> ANSI.default_background()
+        res -> String.to_atom("#{Atom.to_string(res)}_background")
+      end
+    end
+
+    defp without_newline(true), do: ""
+    defp without_newline(false), do: "\n"
   end
 end